Clicking tiles and generation implemented
Title explains it all. All there is to be done now is displaying the nearby bomb count on empty tiles and making the actual gameplay mechanics like the water bar and click count. And main menu, winning and losing, etc. There's more to be done than I thought when I started writing this 😱