Game ready to be published
The core aspects of the game are all finished, this will be my last commit before submitting and packaging the game. Consider this V1.0 as there are still a few things I may feel inclined to add such as particles on click and a main menu with some other game modes and such.